requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0 